Initial modal refactors
