+++ src/components/Modal/utils.js
@@ -1,0 +1,7 @@
+export const COMPONENT_KEY = {
+  Modal: 'Modal',
+  Header: 'Modal.Header',
+  Body: 'Modal.Body',
+  Content: 'Modal.Content',
+  Footer: 'Modal.Footer',
+}
