+++ src/components/Toolbar/types.js
@@ -1,0 +1,14 @@
+export type ToolbarSize =
+| 'xl'
+| 'lg'
+| 'md'
+| 'sm'
+| 'xs'
+
+export type ToolbarTheme =
+| 'default'
+| 'note'
+
+export type ToolbarPlacement =
+| 'top'
+| 'bottom'
